refactor Components names
