+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import './scss/main.scss';
 import App from './js/App';
 import * as serviceWorker from './serviceWorker';
-import Firebase, {FirebaseContext} from "./js/components/Firebase/indexFirebase";
+import Firebase, {FirebaseContext} from "./js/components/Firebase/IndexFirebase";
 
 ReactDOM.render(
   <React.StrictMode>
