+++ src/js/components/Firebase/IndexFirebase.js
@@ -1,0 +1,6 @@
+import FirebaseContext from './Context';
+import Firebase from './Firebase';
+
+export default Firebase;
+
+export { FirebaseContext };
