--- src/js/components/Firebase/indexFirebase.js
@@ -1,6 +1,0 @@
-import FirebaseContext from './context';
-import Firebase from './firebase';
-
-export default Firebase;
-
-export { FirebaseContext };
