+++ src/js/components/Session/withAuthorization.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {withRouter} from 'react-router-dom';
-import {withFirebase} from '../Firebase/context';
+import {withFirebase} from '../Firebase/Context';
 import AuthUserContext from "./context";
 
 const withAuthorization = condition => Component => {
