+++ src/js/components/logic_components/LogOutButton.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { withFirebase } from "../Firebase/context";
+import { withFirebase } from "../Firebase/Context";
 import {Link} from "react-router-dom";
 
 const LogOutButton = ({firebase})=>(
