+++ src/js/landing_page/Discover_section.js
