+++ src/js/landing_page/Footer.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Logo from "../components/logo";
+import Logo from "../components/Logo";
 
 function Footer() {
  return (
