+++ src/js/landing_page/Header.js
@@ -1,5 +1,5 @@
 import React, {useState} from 'react';
-import Logo from '../components/logo'
+import Logo from '../components/Logo'
 import {Link} from "react-router-dom";
 import AuthUserContext from "../components/Session/context";
